backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.10.4
python-dotenv==1.0.1
jsonschema==4.21.1
pytest==8.1.0
pytest-asyncio==0.23.7
mypy==1.8.0
ruff==0.3.0
black==24.2.0
PyYAML==6.0.1
redis==5.0.3
